pyproject.toml
[tool.poetry]
name = "hub-python-sdk"
version = "0.1.0"
description = ""
authors = ["davidhieber <david.hieber@uni-tuebingen.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.7"
yarl = "^1.9.4"
aiohttp = "^3.9.1"
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.8"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"